The correction of the reduced phase: komega_red
The original code was developed with very initial assumption of slowly varying enevelope. THe CEP was not deatl correctly there. Considering the leboratory or co-propagating frame, both the frame-velocity and the CEP phase were not consistently accounted (under the inaccuracy of the SVEA). The actual implementation allows to go to sub-cycle precision. It should be now corrected to treat this dynamics correctly.