Fixed over-allocation and cutoff calculation added
We arrived to the possibility that CUPRAD over-allocates the array for outputs at the beginning as we made this as a sefety factor. Python then should read only slice corresponding to zgrig, that is allways of the right size. The second edit just adds the HHG cut-off as an output analysis.